encagedApe
First release to replace Atom with Brackets. This change is due to the
fact that Atom doesn't play nicely with Cygwin when it comes to locating
executables for its linter and other extensions. It is a separate branch
so that users have a choice about whether or not to upgrade.